change page bug fix
